+++ online-consultant/app/Http/Controllers/Admin/WidgetController.php
@@ -5,8 +5,10 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Requests\Widgets\StoreWidget;
 use App\Http\Requests\Widgets\UpdateWidget;
 use App\Models\Widget;
+use App\Policies\Abilities;
 use App\Services\Companies\CompanyService;
 use App\Services\Widgets\WidgetService;
+use App\Traits\Auth\HasAuthorizationPolicy;
 use Illuminate\Contracts\View\Factory;
 use Illuminate\Http\RedirectResponse;
 use App\Http\Controllers\Controller;
@@ -14,6 +16,10 @@ use Illuminate\View\View;
 
 class WidgetController extends Controller
 {
+    use HasAuthorizationPolicy;
+    
+    protected $modelClass = Widget::class;
+    
     private $widgetService;
     private $companyService;
     
@@ -23,15 +29,21 @@ class WidgetController extends Controller
     ) {
         $this->widgetService = $widgetService;
         $this->companyService = $companyService;
+    
+        $this->viewShareData();
     }
     
     /**
      * Display a listing of the resource.
      *
-     * @return Factory|View
+     * @return Factory|RedirectResponse|View
      */
     public function index()
     {
+        if (!$this->authorizeUserAbility(Abilities::VIEW_ANY)) {
+            return $this->redirectIfNoPermission('admin.dashboard', Abilities::VIEW_ANY);
+        }
+        
         $widgets = $this->widgetService->paginateWidgetsWithTrashed();
         
         return view('admin.models.widgets.index', compact('widgets'));
@@ -40,13 +52,18 @@ class WidgetController extends Controller
     /**
      * Show the form for creating a new resource.
      *
-     * @return Factory|View
+     * @return Factory|RedirectResponse|View
      */
     public function create()
     {
-        $companiesSelectList = $this->companyService->getFormSelectCompanies();
+        if (!$this->authorizeUserAbility(Abilities::CREATE)) {
+            return $this->redirectIfNoPermission('admin.widgets.index', Abilities::CREATE);
+        }
         
-        return view('admin.models.widgets.create', compact('companiesSelectList'));
+        $companiesSelectList = $this->companyService->getFormSelectCompanies();
+        $currentUser = $this->getCurrentUser();
+        
+        return view('admin.models.widgets.create', compact('companiesSelectList', 'currentUser'));
     }
     
     /**
@@ -58,6 +75,10 @@ class WidgetController extends Controller
      */
     public function store(StoreWidget $request): RedirectResponse
     {
+        if (!$this->authorizeUserAbility(Abilities::CREATE)) {
+            return $this->redirectIfNoPermission('admin.widgets.index', Abilities::CREATE);
+        }
+        
         $this->widgetService->createWidget($request->all());
         
         return redirect()->route('admin.widgets.index');
@@ -68,10 +89,14 @@ class WidgetController extends Controller
      *
      * @param  Widget  $widget
      *
-     * @return Factory|View
+     * @return Factory|RedirectResponse|View
      */
     public function edit(Widget $widget)
     {
+        if (!$this->authorizeUserAbility(Abilities::UPDATE, $widget)) {
+            return $this->redirectIfNoPermission('admin.widgets.index', Abilities::UPDATE);
+        }
+        
         return view('admin.models.widgets.edit', compact('widget'));
     }
     
@@ -85,6 +110,10 @@ class WidgetController extends Controller
      */
     public function update(UpdateWidget $request, Widget $widget): RedirectResponse
     {
+        if (!$this->authorizeUserAbility(Abilities::UPDATE, $widget)) {
+            return $this->redirectIfNoPermission('admin.widgets.index', Abilities::UPDATE);
+        }
+        
         $this->widgetService->updateWidget($widget, $request->all());
         
         return redirect()->route('admin.widgets.edit', compact('widget'));
@@ -100,6 +129,10 @@ class WidgetController extends Controller
      */
     public function destroy(Widget $widget): RedirectResponse
     {
+        if (!$this->authorizeUserAbility(Abilities::DELETE, $widget)) {
+            return $this->redirectIfNoPermission('admin.widgets.index', Abilities::DELETE);
+        }
+        
         $this->widgetService->deleteWidget($widget);
         
         return redirect()->route('admin.widgets.index');
@@ -114,7 +147,17 @@ class WidgetController extends Controller
      */
     public function restore(int $id): RedirectResponse
     {
-        $this->widgetService->restoreWidget($id);
+        $widget = $this->widgetService->findWidgetWithTrashed($id);
+    
+        if (!$widget) {
+            return redirect()->route('admin.widgets.index')->with('errors', __('admin.widgets.errors.not_found'));
+        }
+        
+        if (!$this->authorizeUserAbility(Abilities::RESTORE, $widget)) {
+            return $this->redirectIfNoPermission('admin.widgets.index', Abilities::RESTORE);
+        }
+        
+        $this->widgetService->restoreWidget($widget);
         
         return redirect()->route('admin.widgets.index');
     }
@@ -128,7 +171,17 @@ class WidgetController extends Controller
      */
     public function forceDelete(int $id): RedirectResponse
     {
-        $this->widgetService->forceDeleteWidget($id);
+        $widget = $this->widgetService->findWidgetWithTrashed($id);
+    
+        if (!$widget) {
+            return redirect()->route('admin.widgets.index')->with('errors', __('admin.widgets.errors.not_found'));
+        }
+        
+        if (!$this->authorizeUserAbility(Abilities::FORCE_DELETE, $widget)) {
+            return $this->redirectIfNoPermission('admin.widgets.index', Abilities::FORCE_DELETE);
+        }
+        
+        $this->widgetService->forceDeleteWidget($widget);
         
         return redirect()->route('admin.widgets.index');
     }
